+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - fix using shared interface type in multiple schemas
 - fix calling field resolver without providing resolver class to `buildSchema`
 - fix generated TS union type for union type of object type classes extending themselves (#587)
+- fix using shared union and interface types in multiple schemas when `resolveType` is used
 ### Others
 - **Breaking Change**: change build config to ES2018 - drop support for Node.js < 10.3
 - **Breaking Change**: remove deprecated `DepreciationOptions` interface